requirements.txt
beautifulsoup4==4.12.2
chainlit>=1.1.306
chromadb==0.5.0
injectable==4.0.1
langchain==0.2.5
langchain-experimental==0.0.61
langchain-core==0.2.9
langchain-community==0.2.5
langchainhub==0.1.20
langserve>=0.0.51
langsmith>=0.1.17,<0.2.0
openai>=1.26.0,<2.0.0
tqdm==4.66.1
langchain_openai==0.1.8
langchain-text-splitters==0.2.1
pandas==2.0.3
ttkthemes==3.2.2
PyQt5==5.15.9
unstructured==0.7.7
markdown==3.4.4
psutil==5.9.5
flask==2.3.3
fastapi>=0.110.1,<0.113
uvicorn>=0.25.0,<0.26.0 
python-dotenv>=1.0.0
onnxruntime==1.18.1
pydantic==2.11.5
langfuse==2.60.8

